pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "homer-redaction-detector"
version = "2.0.0"
description = "Detect hidden text under PDF redaction boxes"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "PyMuPDF>=1.23,<2.0",
]

[project.optional-dependencies]
ocr = [
    "Pillow>=10.0,<12.0",
    "pytesseract>=0.3.10,<1.0",
]
fast = [
    "numpy>=1.24,<3.0",
]
repair = [
    "pikepdf>=8.0,<10.0",
]
all = [
    "homer-redaction-detector[ocr,fast,repair]",
]

[project.scripts]
homer = "homer.__main__:main"

[tool.setuptools.packages.find]
include = ["homer*"]
